+++ packages/hooks/demo/App.tsx
@@ -1,4 +1,4 @@
-import '@cutting/component-library/dist/component-library.cjs.development.css';
+import '@cutting/component-library/dist/component-library.cjs.production.min.css';
 import { FC, useRef } from 'react';
 import { useParentSize } from '../src/useParentSize/useParentSize';
 import { ApplicationLayout } from '@cutting/component-library';