+++ packages/svg/demo/App.tsx
@@ -1,4 +1,4 @@
-import '@cutting/component-library/dist/component-library.cjs.development.css';
+import '@cutting/component-library/dist/component-library.cjs.production.min.css';
 import type { FC } from 'react';
 import { useRef } from 'react';
 import { ApplicationLayout } from '@cutting/component-library';