+++ packages/website/src/components/OSS/repos.ts
@@ -8,17 +8,17 @@ export const repos: Repo[] = [
   {
     name: '@cutting/react-abortable-fetch',
     description: 'The useQuery I needed',
-    link: 'https://github.com/dagda1/cuttingedge/tree/master/packages/react-abortable-fetch',
+    link: 'https://github.com/dagda1/cuttingedge/tree/main/packages/react-abortable-fetch',
   },
   {
     name: 'use-shortcuts',
     description: 'Effortlessly create keyboard shortcuts with this react hook.',
-    link: 'https://github.com/dagda1/cuttingedge/tree/master/packages/use-shortcuts',
+    link: 'https://github.com/dagda1/cuttingedge/tree/main/packages/use-shortcuts',
   },
   {
     name: 'use-mathjax',
     description: 'easy mathjax in react',
-    link: 'https://github.com/dagda1/cuttingedge/tree/master/packages/use-mathjax',
+    link: 'https://github.com/dagda1/cuttingedge/tree/main/packages/use-mathjax',
   },
   {
     name: 'bigtest',