+++ packages/website/src/containers/App/index.tsx
@@ -1,4 +1,4 @@
-import '@cutting/component-library/dist/component-library.cjs.development.css';
+import '@cutting/component-library/dist/component-library.cjs.production.min.css';
 import { FC, StrictMode } from 'react';
 import { Routes } from '../../routes';
 import { BrowserRouter as Router } from 'react-router-dom';