+++ packages/component-library/CHANGELOG.md
@@ -1,6 +1,16 @@
 # @cutting/component-library
 
+## 4.23.3
+
+### Patch Changes
+
+- Fix babel issues and include a umd build for browser
+- Updated dependencies [undefined]
+  - @cutting/hooks@4.24.2
+  - @cutting/util@4.24.1
+
 ## 4.23.2
+
 ### Patch Changes
 
 - Updated dependencies [undefined]