+++ packages/graphql-explorer/CHANGELOG.md
@@ -1,6 +1,16 @@
 # @cutting/graphql-explorer
 
+## 4.22.5
+
+### Patch Changes
+
+- Fix babel issues and include a umd build for browser
+- Updated dependencies [undefined]
+  - @cutting/react-abortable-fetch@0.9.4
+  - @cutting/util@4.24.1
+
 ## 4.22.4
+
 ### Patch Changes
 
 - Updated dependencies [undefined]