+++ packages/hooks/CHANGELOG.md
@@ -1,6 +1,15 @@
 # @cutting/hooks
 
+## 4.24.2
+
+### Patch Changes
+
+- Fix babel issues and include a umd build for browser
+- Updated dependencies [undefined]
+  - @cutting/util@4.24.1
+
 ## 4.24.1
+
 ### Patch Changes
 
 - Updated dependencies [undefined]