+++ packages/react-abortable-fetch/CHANGELOG.md
@@ -1,6 +1,15 @@
 # @cutting/react-abortable-fetch
 
+## 0.9.4
+
+### Patch Changes
+
+- Fix babel issues and include a umd build for browser
+- Updated dependencies [undefined]
+  - @cutting/util@4.24.1
+
 ## 0.9.3
+
 ### Patch Changes
 
 - Updated dependencies [undefined]