+++ packages/svg/CHANGELOG.md
@@ -1,6 +1,16 @@
 # @cutting/svg
 
+## 4.24.1
+
+### Patch Changes
+
+- Fix babel issues and include a umd build for browser
+- Updated dependencies [undefined]
+  - @cutting/hooks@4.24.2
+  - @cutting/util@4.24.1
+
 ## 4.24.0
+
 ### Minor Changes
 
 - Add nested Mathax svg support