+++ packages/use-mathjax/CHANGELOG.md
@@ -1,6 +1,15 @@
 # @cutting/use-mathjax
 
+## 4.24.1
+
+### Patch Changes
+
+- Fix babel issues and include a umd build for browser
+- Updated dependencies [undefined]
+  - @cutting/util@4.24.1
+
 ## 4.24.0
+
 ### Minor Changes
 
 - Add nested Mathax svg support