+++ packages/util/CHANGELOG.md
@@ -1,6 +1,13 @@
 # @cutting/util
 
+## 4.24.1
+
+### Patch Changes
+
+- Fix babel issues and include a umd build for browser
+
 ## 4.24.0
+
 ### Minor Changes
 
 - Add nested Mathax svg support