+++ packages/website/CHANGELOG.md
@@ -1,6 +1,20 @@
 # @cutting/website
 
+## 4.24.1
+
+### Patch Changes
+
+- Fix babel issues and include a umd build for browser
+- Updated dependencies [undefined]
+  - @cutting/hooks@4.24.2
+  - @cutting/react-abortable-fetch@0.9.4
+  - @cutting/svg@4.24.1
+  - @cutting/component-library@4.23.3
+  - @cutting/use-mathjax@4.24.1
+  - @cutting/util@4.24.1
+
 ## 4.24.0
+
 ### Minor Changes
 
 - Add nested Mathax svg support